Moved software testing description to appendix
The software testing methodology (test driven development or TDD) is
moved to the appendix, since it is a general best practice and not
related to the specifics of verifying the functionality an obstacle
detection system.

The precise steps we will take to ensure our software works need to be
outlined.

Added references for TDD.